extract currency layer factory method
