+++ src/store/handler/currencyLayer/handler.go
@@ -1,21 +1,21 @@
-package handler
+package currencyLayer
 
 import (
 	"github.com/voyago/converter/environment"
 	"github.com/voyago/converter/src/model"
 )
 
-type CurrencyLayer struct {
+type Handler struct {
 	Currency string
 	Env environment.Env
 }
 
-func (current CurrencyLayer) ExchangeRates() model.Currencies {
+func (current Handler) ExchangeRates() model.Currencies {
 	items := make(map[string]model.Currency)
 
 	return model.Currencies{Items: &items}
 }
 
-func (current CurrencyLayer) ApiKey() string  {
+func (current Handler) ApiKey() string  {
 	return current.Env.Get("CONVERTER_CURRENCY_LAYER_KEY")
 }
