+++ tests/unit/store/store_test.go
@@ -2,7 +2,7 @@ package store
 
 import (
 	"github.com/voyago/converter/src/store"
-	"github.com/voyago/converter/src/store/handler"
+	"github.com/voyago/converter/src/store/handler/currencyLayer"
 	"testing"
 )
 
@@ -31,19 +31,15 @@ func TestItProperlyBuildsTheCurrencyLayerDriver(t *testing.T) {
 		t.Errorf("The given store base currency is invalid")
 	}
 
-	assert := func(current handler.CurrencyLayer) {
-		if current.ApiKey() != "" {
-			t.Errorf("The given key is invalid, %s", current.ApiKey())
+	switch value := manager.Handler.(type) {
+	case currencyLayer.Mock:
+		if value.ApiKey() != "" {
+			t.Errorf("The given key is invalid, %s", value.ApiKey())
 		}
 
-		if current.Currency != "SGD" {
+		if value.Currency != "SGD" {
 			t.Errorf("The given currency is invalid")
 		}
-	}
-
-	switch value := manager.Handler.(type) {
-	case handler.CurrencyLayer:
-		assert(value)
 	default:
 		t.Errorf("The given store handler is invalid")
 		t.FailNow()
